rename Mod_Name SCF as scf_method
